DEV-2752 Fix event time as int
There was an issue with the refactor. The class "datetime" is imported
instead of the module but the code was not adapted properly.

Also, the occurred-at timestamp is a timezone-aware timestamp. The epoch
timestamp (datetime.utcfromtimestamp()) however is timezone-naive. It is
not possible to subtract a naive timestamp from an aware one: "TypeError:
can't subtract offset-naive and offset-aware datetimes". Changed the epoch
timestamp so that it reuses (!) the tzinfo from the occurred-at timestamp.
The epoch timestamp can be timezone-aware or timestamp-naive, depending one
occurred-at timestamp.

Finally, the test imported from the bytecode file (.pyc) and not the actual
source code (.py). Fixed the import to point to the source code.